pyproject.toml
[tool.poetry]
name = "basic-django"
version = "0.1.0"
description = ""
authors = ["2280250608 <h.izadyfar@gmail.com>"]
readme = "README.md"
packages = [{include = "basic_django"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
jwt = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
